requirements.txt
Flask==2.1
names==0.3.0
faker==13.11.1
mysql--connector-python==8.0.29